+++ CHANGELOG.md
@@ -21,6 +21,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 --- -->
 
+## [0.2.2] - 2026-03-02
+
+### Added
+
+- Add FieldSelector component and integrate with Header and NavMenu for managing visible fields
+
+---
+
 ## [0.2.1] - 2026-03-02
 
 ### Changed